+++ include/inc_lang/backend/de/lang.inc.php
@@ -1298,6 +1298,7 @@ $BL['fancyupload_flash_error']			= 'Installieren Sie das neueste Adobe Flash Plu
 
 $BL['be_cnt_function_validate']			= 'PHP-Validierung';
 $BL['be_structform_selected_cp']		= 'Auswahl nutzbarer Content Parts begrenzen';
+$BL['be_structform_select_cp']			= 'Content Parts w&auml;hlen';
 
 $BL['source_image_not_found']			= 'Quelldatei Fehler: Die Datei %s konnte nicht gefunden werden.';
 $BL['form_force_ssl']					= 'Formularversand per SSL erzwingen';