+++ include/inc_lang/backend/en/lang.inc.php
@@ -1308,6 +1308,7 @@ $BL['fancyupload_flash_error']			= 'To enable the embedded uploader, install the
 
 $BL['be_cnt_function_validate']			= 'PHP validation';
 $BL['be_structform_selected_cp']		= 'Limit selection of usable content parts';
+$BL['be_structform_select_cp']			= 'Select content parts';
 
 $BL['source_image_not_found']			= 'Source image error: The image %s seems not to exist';
 $BL['form_force_ssl']					= 'Force sending forms with SSL';